flatten some folder stuff
