+++ web-app/src/app/app-shell/PwaInstaller.tsx
