+++ web-app/src/app/design-system/Icons/PWA.tsx
