+++ web-app/src/app/design-system/Icons/Robot.tsx
