+++ web-app/src/app/design-system/Icons/index.ts
@@ -1,2 +1,3 @@
 export * from "./Robot";
 export * from "./PWA";
+export * from "./WindyCivi";
