+++ web-app/src/app/design-system/Icons/utils.tsx
