+++ web-app/src/app/design-system/Icons/windy-civi-logo.png
