+++ web-app/src/app/design-system/styles/enums.ts
@@ -1,11 +1,6 @@
-export enum MediaQueries {
-  "sm" = "media (min-width: 640px)",
-  "md" = "media (min-width: 768px)",
-  "lg" = "media (min-width: 1024px)",
-  "xl" = "media (min-width: 1280px)",
-  "2xl" = "media (min-width: 1536px)",
-}
-
+/**
+ * Used in style.d.ts to define the types of the styles
+ */
 export enum Spacing {
   ZERO = "0",
   ONE = "5px",
