+++ web-app/src/app/design-system/styles/types.ts
