+++ web-app/src/app/feed-ui/react/Representatives.tsx
@@ -3,7 +3,7 @@ import type { OfficialOffice } from "@windy-civi/domain/representatives/represen
 import { getLegislators } from "@windy-civi/domain/representatives/representatives.utils";
 import { LocationFilter } from "@windy-civi/domain/types";
 import { useState, type FC } from "react";
-import { DataField, ResultCard } from "~app/modules/design-system";
+import { DataField, ResultCard } from "../../design-system";
 import { FeedFilterProps } from "../feed-ui.types";
 
 export const RepresentativesList: FC<{
