--- web-app/src/app/modules/design-system/Instructions/Instructions.tsx
@@ -1,8 +1,0 @@
-export function Instructions() {
-  return (
-    <div className="grow p-6">
-      <h2 className="text-xl font-bold">Find Your Representatives</h2>
-      <p>Use this site to find information about your representatives.</p>
-    </div>
-  );
-}
