--- web-app/src/app/modules/design-system/index.ts
@@ -1,13 +1,0 @@
-export * from "./AddressLookup/AddressLookup";
-export * from "./Button/Button";
-export * from "./Card/Card";
-export * from "./DataField/DataField";
-export * from "./Divider/Divider";
-export * from "./Instructions/Instructions";
-export * from "./Layout/Layout";
-export * from "./Logo/Logo";
-export * from "./Modal/Modal";
-export * from "./RadioPicker/RadioPicker";
-export * from "./ResultCard/ResultCard";
-export * from "./styles";
-export * from "./Tags/Tags";
